core: send discard to vdsm according to DAD value
This patch adds a new parameter, "discard", to all 5 relevant VDS
commands: CopyImage, DeleteImage, DestroyImage, MergeSnapshots and
MoveImageGroup.
Each VDS command is given with the new parameter according to the value
of "Discard After Delete" of the relevent storage domain.

Note that since only 4.1 engines and vdsms are updated with the new api,
in vdsm the parameter has a default value of false, and old engines will
still call the old api. Therefore, old engines can work with new vdsms
and so for new engines and old vdsms.

Change-Id: I0d4239816927d313674553fbfa5c1744c5c54174
Bug-Url: https://bugzilla.redhat.com/1342919
Signed-off-by: Idan Shaby <ishaby@redhat.com>