webadmin: Chaning default cluster attachment for external networks
Until now the external networks were by default not attached to any cluster
when creating, nor was it possible to attach it.
This was very confusing for users.

This patch changes this, attaching an external network to all clusters,
when an external network is created.
It is now also possible to attach/unattach the network from/to clusters.
Enabling attaching/unattaching is important as not all the clusters
could be configured for a given external network.

The 'required' attribute for a network on a cluster stays disabled for
the external network, as an external network can not be a required network.

Whenever a change to the external status changes, we need to set all the
cluster's isAttached values again, in case the user has modified any of
them.

Change-Id: Ic335c4f5bf53da5ce0c3dd48b5e807950a801c16
Bug-Url: https://bugzilla.redhat.com/1373477
Bug-Url: https://bugzilla.redhat.com/1383449
Signed-off-by: mirecki <mmirecki@redhat.com>