core: StorageJob - adding support to inspect the job error code
When a storage job fails with some error we may want to perform some
operations or to consider the execution as succesful.
For example - if a deletion fails because the storage object doesn't
exist the engine may want to consider the operation as succesful.

This patch adds support to handle the error code of failed jobs and to
return a different status if needed.

Change-Id: Ifad25682f29e0fcda532aa0f41cb2a50e02cfb73
Bug-Url: http://bugzilla.redhat.com/1403839
Signed-off-by: Liron Aravot <laravot@redhat.com>