engine: various fixes in sysprep XML templates
The commit performs several commits:

1) Reorder elements in the template
The order of elements is arbitrary to some degree. The commit reorders
the elements so that the order is same between 32-bit and 64-bit version
of the XML. This leads to nicer diffs and makes it easier to spot
errors. In the process some useless comments were removed.

2) Change x86 sections to amd64 in 64-bit templates
Each component block in the sysprep template is tied to the
architecture. We have had mistakenly some of the blocks marked as x86 in
templates for 64-bit systems.

3) Fix messed up locales in sysprep templates
There were following problems with locales:
- some locales were expanded from wrong variables
- some locales had value en-US hardcoded
- some input locales used hardcoded 0409:00000409 (that is US keyboard)

Bug-Url: https://bugzilla.redhat.com/1402817
Change-Id: I33840d74d157efddeb280a52318517d8b15db29f
Signed-off-by: Tomáš Golembiovský <tgolembi@redhat.com>