engine: reorder elements in sysprep XML templates
The order of elements is arbitrary to some degree. The commit reorders
the elements so that the order is same between 32-bit and 64-bit version
of the XML. This leads to nicer diffs and makes it easier to spot
errors.

In the process some useless comments were removed. Apart from this the
commit does not introduce any changes to the elements themselves.

Change-Id: Ide66a00cc64a653828a8ca4d376e139811aa6574
Signed-off-by: Tomáš Golembiovský <tgolembi@redhat.com>