core: introducing ReduceSANStorageDomainDevicesCommand
This patch adds the EditSANStorageDomainDevicesCommand that'll be used
to edit devices of a block storage domain. The command currently supports
removing devices from domains and may be extended in the future to
support domain extension as well.

The command recieves as a parameter the  list of devices to remove from the
domain, the command will remove the devices one by one - making sure the data
from each device is moved to devices that aren't in the removal list.

Change-Id: I44e905833ecac7a6eba20bb418db82a3a622c4e4
Bug-Url: http://bugzilla.redhat.com/1403839
Signed-off-by: Liron Aravot <laravot@redhat.com>