core: add vm lease on add/edit/import vm
Add VM lease in the following flows:

Add VM - if we fail to create the lease the VM is not added.
Import VM - if we fail to create the lease, the VM is imported without lease.
Edit VM - if we fail to create the lease, the VM is not updated. Note that
even if we fail to remove the previous lease, we proceed with the update.

Change-Id: I7c666a17d664c2b73ea7cfef10987f12b2bc1b5c
Signed-off-by: Arik Hadas <ahadas@redhat.com>