core: a more elaborate audit log for management network on broken bond
the designated interface for network management may be filtered out (by
filterBondsWithoutSlaves) as a result of being an incorrectly configured
bond (consisting of less than 2 active slaves).

filterBondsWithoutSlaves returns an array of interfaces that are
either appropriate bonds (=> 2 active slaves) or non-bonds. if the 
designated interface is filtered out as a result, it is due to being an 
incorrect bond.

Change-Id: I6b6495a1d1147d9364317001f568bc9a9384d17c
Bug-Id: https://bugzilla.redhat.com/1368771
Signed-off-by: Leon Goldberg <lgoldber@redhat.com>