core: added error log for illegal host status during network setup
when trying to setup a network while host status isn't compatible,
an error message is sent, however it fails to include the status of the
host; it only mentions the allowed.

as of now we're only interested in knowing the status internally for
debugging purposes and so a simple log message was added.

Change-Id: I6c227607cfafbbd5e493c709e1484c1001c10607
Signed-off-by: Leon Goldberg <lgoldber@redhat.com>