core: private VmCommand#cdPathWindowsToLinux(String, String)
This method is an internal implementation detail of
VmCommand#cdPathWindowsToLinux(String, Guid, Guid), and as such
should not be used in other contexts.

This patch cleans up RunVmCommandTest to mock
getIsoPrefix(Guid, Guid) directly instead of setting up a convoluted
doAnswer behavior and changes this method to be private, as it ought
to be.

Change-Id: I643352a4d00afdb62d219f372b98a25b513d1709
Signed-off-by: Allon Mureinik <amureini@redhat.com>