core: remove redundant validation in create-vm
We usually do not validate things in vds commands since they are
supposed to be validated in the bll commands. We have some
legacy checks in CreateVm though. This patch removes one of them
since we need the ability to run VMs in UNKNOWN status (these
that are configured with a VM lease only).

Change-Id: I0cbfe7ae1a2b144577c37a77656a5b4a251075a6
Signed-off-by: Arik Hadas <ahadas@redhat.com>