core: dropping sd_static luns fk
Currently storage_domain_static has FKs to the luns table for its
metadata devices (relevant for block domains).
Some of the flows remove the lun before the sd, causing to FK
violation. As a quick fix, the FK is removed and we'll be able to
restore it after when those flows are fixed.

Change-Id: I5df3da87db06d7f9fb2cb48eb1d3af1d1aea3f2d
Signed-off-by: Liron Aravot <laravot@redhat.com>