webadmin: Correcting height in VfsConfigPopupView.
Since the relative height seems not to work like
expected, the internal height of a panel is hard coded.
This causes the vertical scrollbar is shown if needed.
To gain horizontal space for the scrollbar, the width of
the panel and the popup view is extended.

Change-Id: I017711d70444d7058da159654fedb798af597bd3
Bug-Url: https://bugzilla.redhat.com/1341162
Signed-off-by: Dominik Holler <dholler@redhat.com>
(cherry picked from commit d264d640fab499e2e81fcbbd6c738100b5e1e000)