core: SyncLunsInfoForBlockStorageDomainCommand#cleanupLunsFromDb
Improved the method performance to collect the IDs of the LUNs from the
VG to a single Set once and then use it for searching existing LUNs.
Given M luns from the VG and N luns in the DB, this technique does O(N)
comparisons instead of O(M*N).

Change-Id: Id7dfc76cb7781d6286161b050b4093411269682d
Signed-off-by: Allon Mureinik <amureini@redhat.com>