core: fix NPE when compensating after engine shutdown
When the engine shuts down during attachment of an SD
to a DC, the compensation mechanism will fail because
of an NPE when the engine is restarted.

This is caused because businessEntitySnapshotDao in
CommandBase is null when CompensationContext#afterComepnsationCleanup
is called.

Fixed by moving the creation of DefaultCompensationContext to
CommandBase#postConstruct so it executes after the dependencies have
been injected.

Change-Id: I33866ebf866926f34e39d104151977eee948fe88
Bug-Url: https://bugzilla.redhat.com/1404727
Signed-off-by: Benny Zlotnik <bzlotnik@redhat.com>