core: Populate V3 product_info.version.full_version
In version 3 of the API the full version of the product is potentially
available in two places:

  product_info.version.full_version
  product_info.full_version

Version 3 of the engine used to populate the second one, but the version
3 compatibility layer of version 4 of the engine populates only the
second. This means that clients that still use version 3 of the API
won't be able to get this information. To avoid that, this patch changes
the V3 compatibility layer so that it populates both places with the
same value.

Change-Id: If8a60ba2ef333dc87cc113959f5bdb3f35426b09
Bug-Url: https://bugzilla.redhat.com/1405386
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>
(cherry picked from commit 21ef0fba8779f51a10b6e811bfc55504353f590d)
(cherry picked from commit 33601390074ea1ea7c6b743272db521f793070fa)