packaging: setup: More aggressive autovacuum configuration
Configuring the autovacuum daemon to be more performant and aggressive.
To make the daemon efficient we a) tell it to start working more frequent
on small number of changes (vacuum_scale_factor), b) spin more workers
as we have more then 3 tables which are updated frequently and we want
to prevent from exhausting the vacuum cost on few tables and starving
other tables from being vacuumed and c) raising the memory amount each
worker have to the passes on the pages would be efficient and quick.

Our v*_dynamic and v*_statistics tables (plus few more) are updated very
frequently as part of our monitoring needs. Even though we keep on
optimizing the amount of updates to the bare minimum we still have to
make sure we don't starve any of those tables otherwise we are in the
risk of bloating the db.

Change-Id: I426de8d2903dc8430e14957738e04757eb4c7eb5
Bug-Url: https://bugzilla.redhat.com/1388433
Signed-off-by: Roy Golan <rgolan@redhat.com>