core: remove unneeded query when getting vms to move to unknown
When host monitoring gets a network exception during monitoring of a
specific host, it switches this host to maintenance and moves all the
VMs running on this host to Unknown.

The retrieval of VMs that need to be switched to unknown was composed
of two parts:
1. get all the VMs running on the host (their run_on_vds points to
that host).
2. get all the VMs that migrate to the host (their migrating_to_vds
points to that host) and also are in MigratingTo status.

But (2) is redundant since at this phase of the migration run_on_vds
must point to the destination host. Therefore, this part is removed.

Note: the query for VM dynamics from the database was moved as a
preparation for future changes.

Change-Id: I1727da9d56cf3b21464da913942f115abeb1bd5b
Signed-off-by: Arik Hadas <ahadas@redhat.com>