core: auto start vms with lease
When detecting that a host is non-responsive we move the VMs that
are supposed to run on that host to Unknown status. This patch
adds to that a special treatment for VMs with leases: for each
VM with a lease, we will try to run the VM on another host.
The lease should prevent us from getting a split-brain. However,
we do not want to try to run the VM on each vds monitoring cycle
(which is by default every 3 seconds) so we define a config value
for setting the frequency of those attempts.

Note that when host becomes non-responsive we get network exceptions
from several places:
1. calls to refresh host capabilities.
2. calls to poll vms (on statistics iterations).
3. every command that interacts with vdsm.

As a result, VdsManager#restartVmsWithLeaseIfNeeded may be called
simultaneously from several threads and that may cause us to skip
a restart attempt. Thus, adding a lock to ensure mutually exclusive
execution of that method.

Change-Id: I464f14fe953a6fedfa53d6655eeb31965b5f1579
Bug-Url: https://bugzilla.redhat.com/1317429
Signed-off-by: Arik Hadas <ahadas@redhat.com>