engine-backup: using a flag to track a restored DB
Introducing a new vdcOption flag to track a DB
that has been just restored from a backup.
On a restored DB not all the assumptions are
alwayws valid; for instance we cannot check
HE global maintenance mode there since it
will just report the status at DB backup
creation time and not the actual status.

Change-Id: I8e99b3f2240f157b72051accc96f0114333ffcbd
Bug-Url: https://bugzilla.redhat.com/1403903
Signed-off-by: Simone Tiraboschi <stirabos@redhat.com>