core: DeactivateStorageDomainWithOvfUpdate - context passing
Currently DeactivateStorageDomainWithOvfUpdate passes its compensation
context to the executed DeactivateStorageDomain command.

When passing the compensation context along the hirerchy, it was decided
in BZ 1364804 that the reponsiblity to revert the compensation data is on
the parent command.

DeactivateStorageDomainWithOvfUpdate is always succesfull (so is its
end methods) - and therefore in case of error in the child command
execution the compensation context isn't reverted which causes the domain
to remain locked.
This patch solves that by not passing the context to the executed
command - so it will have a compensation context of its own.

Change-Id: Iebbb3084d52c7744d1f92a98c31d3a3c916e6e03
Bug-Url: https://bugzilla.redhat.com/1405772
Signed-off-by: Liron Aravot <laravot@redhat.com>