webadmin: Correcting size of VfsNicLabelWidget.
The VfsNicLabelWidget takes minimal to much width to be displayed
without a vertical scrollbar, but does not use all vertical space
available.

The width of the widget, but not it's content, is slightly decreased
to avoid the horizontal scrollbar. The height is increased to use the
available space.

Change-Id: If8ee2325163a64cda5bfa557dd2d71d87f6e088d
Bug-Url: https://bugzilla.redhat.com/1406808
Signed-off-by: Dominik Holler <dholler@redhat.com>
(cherry picked from commit ff3d3a08c8d71a665f497a67b0f0d6967a088737)