core: rename and use literal ip in getMigrationNetworkIp
all currently-supported versions of vdsm expect what rfc#2732 calls
"literal ip addresses" as for the dstqemu parameter of the migrate verb.

as part of the ipv6 migration effort, it is required to use ip
addresses that are uri compatible, i.e. ip literals.

the ipv4 variation requires no further modifications, however if
there's no ipv4 and ipv6 is used, complying to rfc#2732 is necessary.

the function was renamed to both indicate and explain the string
modification.

Change-Id: I5e3349f349e737b5198bb5100e76c8f25699e9bf
Signed-off-by: Leon Goldberg <lgoldber@redhat.com>