core: allow ipv6 address for migration role in cluster network
we're now allowing migration networks to consist of interfaces
that solely use ipv6. it is already possible to setup a host migration
attachment to use ipv6, however it is not possible to set migration role
when some host in a cluster doesn't have an ipv4 address in the relevant
nic.

when determining what's considered a missing ip in a nic, check for the
role of the cluster network, and if it's migration, consider it to be
missing only if both ipv6 and ipv4 addresses are missing.

Change-Id: Iae51623ebbfc2431d5f42f1e026c853afbbedb5a
Signed-off-by: Leon Goldberg <lgoldber@redhat.com>