bll: Prevent migration to the same host when balancing
This adds a new command - BalanceVm - that is the same as
MigrateVm in all aspects except one:

It allows the scheduler to select the host the VM is
currently running at and does nothing when it is selected
as the best host. All other results cause an ordinary
migration to be started.

Bug-Url: https://bugzilla.redhat.com/1410040
Change-Id: I466a72a546a048a177c55ea2ca6a1a0b697bd53e
Signed-off-by: Martin Sivak <msivak@redhat.com>