core: do not ignore down event on destroy vm
Since we introduced the separation of host and VMs monitoring plus events,
it takes too much time to power-off a VM. Power-off should be a quick
operation - similar to unplug the power cable from a physical machine.
However, it can now take up to 15 seconds in the worst case for this
operation to finish (and even 20sec considering the UI refresh rate):
1. the engine locks the VM with its monitoring lock.
2. engine sends destroy call to VDSM.
3. VDSM sends an event that the VM is destroyed.
4. the engine ignores the event since because of the VM is locked.
5. the VM remains in Powering Down status until the next polling cycle.

Sure, events should be queued and not ignored, but this is a too complex
change to do at the moment. This patch introduces an alternative approach
that eliminates the lock in destroy-VM flow.

Change-Id: I14931877309867bbbc0115e3a384e39757e3234c
Bug-Url: https://bugzilla.redhat.com/1367023
Signed-off-by: Arik Hadas <ahadas@redhat.com>