scheduling: Fix AREM logic for host affinity
This patch updates the AREM algorithms with regards to host affinity.
We sadly have a logic bug in the previously merged version.

All affinity groups have to be considered separately when a VM is part
of multiple host affinity groups. The correct operation for selecting
a valid host is intersection, not union.

Bug-Url: https://bugzilla.redhat.com/1404660
Change-Id: I683c46ef3be8e94cacbb554c5a372a9574f584d4
Signed-off-by: Martin Sivak <msivak@redhat.com>