scheduling: Fix AREM logic for host affinity
This patch updates the AREM algorithms with regards to host affinity.
We sadly have a logic bug in the previously merged version.
All affinity groups have to be considered separately when a VM is part
of multiple host affinity groups. The correct operation for selecting
a valid host is intersection, not union.

Change-Id: I2c95ae9ece924263accdfba5deac084506a6da2a
Bug-Url: https://bugzilla.redhat.com/1404660
Signed-off-by: Martin Sivak <msivak@redhat.com>
(cherry picked from commit 48e6d23)