restapi: Use non-capturing groups for action status
The API has a mechanism that can be used to check the status of
asynchronous actions. For example, when a template is imported the API
returns a response like this:

  <action href="/ovirt-engine/api/storagedomains/123/templates/456/import/789"/>
    ...
  </action>

The 'href' provided in that response can be used to check the status of
the action.

In the server side that is implemented by methods with declarations like
this:

  @Path("{action: (import|register)}/{oid}")
  ActionResource getActionResource(
    @PathParam("action") String action,
    @PathParam("oid") String oid
  );

But unfortunatelly since Resteasy 3 there is a bug that makes this fail:
the parenthesis don't work correctly with resource locators. See this
for details:

  Regex with parenthesis doesn't work with resource locator
  https://issues.jboss.org/browse/RESTEASY-1569

To avoid that issue this patch changes updates the server to use version
1.1.10 of the metamodel, which generates that code using non-capturing
groups instead of plain parenthesis:

  @Path("{action: (?:import|register)}/{oid}")
  ActionResource getActionResource(
    @PathParam("action") String action,
    @PathParam("oid") String oid
  );

Note the use of (?:...) instead of just (...).

That solves the issue for all the V4 action resources, because the
code is automatically generated by the metamodel. But for the V3
compatibility layer this has to be modified manually in each resource.

In addition the URLs generated by the action resource need to be
generated correctly, containing the "/v3" or "/v4" prefix only when it
is explicitly included in the request.

Change-Id: Ied1d04175706d38d725d4ea05617b0c93ec8ca91
Bug-Url: https://bugzilla.redhat.com/1405774
Related-To: https://issues.jboss.org/browse/RESTEASY-1569
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>