engine: set a valid singleQxlPci value for v2v VMs
In case of v2v from KVM, the singleQxlPci value of the imported VM was
always set to 'true' without any validation (since it isn't read from
OVF in case of KVM, it was set to a default value).
The singleQxlPci is supported only when the VM's OS is Linux
and the Video device type is QXL and only on those cases it should be
set to 'true'.

Since the singleQxlPci validation is done in updateVm command, in case
of a none Linux VM or if video type is not QXL, each update operation
via the APIs failed in validation phase.

This is the exact senario occured in BZ 1408715.

Change-Id: I7c01d632652e5982cf7d87da8f64e57ebbc6f603
Bug-Url: https://bugzilla.redhat.com/1408715
Signed-off-by: Sharon Gratch <sgratch@redhat.com>