backend: fix logging of disks with illegal Pass Discard
Previously, when adding a cloned vm from a template, if at least one of
the target vm's disks had an illegal Pass Discard value, the wrong
disks' ids were logged (the source disks' ids instead of the target
ones).
This patch simplifies the way we log these disks by moving the log
handling to BlockStorageDiscardFunctionalityHelper and fixes this bug,
so that now the right ids are logged.

Change-Id: I566b36d3b7c25c4b8badfa59a9113fd3885eac4d
Bug-Url: https://bugzilla.redhat.com/1410017
Signed-off-by: Idan Shaby <ishaby@redhat.com>