core: ReduceSANStorageDomainDevices - free space validation
The user can execute the ReduceSANStorageDomainDevices command with few
devices to reduce from the domain.
ReduceSANStorageDomainDevices executes RemoveDeviceFromSANStorageDomain
for each of the devices, this command first moves the data from the
device and then reduces the domain by removing the device from it.
The data is moved using the lvm command pvmove, if there's not enough
free space the operation will fail.
When attempting to reduce multiple devices, we can't rely on the lvm
verification for space (as lvm isn't aware that the user actually wants
to reduce more than one device).

This patch adds verification that there's enough free space for the
operation - the command executes getVgInfo, sums the pe allocation of
the devices for reduction and checks if the number of the free extents
is sufficent.

Change-Id: I80cfb95b61692ea83c3eaf2c490d20b17b2d0236
Signed-off-by: Liron Aravot <laravot@redhat.com>