core: Pull up storage domain static and dynamic in fixtures.xml
Since introducing VM leases the VmStatic entity can hold a reference to a
storage domain with the column 'lease_sd_id', in fixtures.xml the
storage_domain_static table is defined after the vm_static table causing
any VM that holds a reference to a storage domain to fail the test
database creation on a foreign key constraint since the
storage_domain_static table is not created yet.
This patch moves the storage_domain_static and storage_domain_dynamic
table up, right after storage_pool to avoid this problem, this is also
more logical given the heirarchy of storage domain in the system.

Change-Id: I508876248ec5867974fe2cf3de4f6ecc17b5c42e
Signed-off-by: Tal Nisan <tnisan@redhat.com>