core: Introducing CreateAllTemplateDisksCommand
Adding CreateAllTemplateDisksCommand that creates all disks needed for
creating a template. This command is called from AddVmTemplateCommand
instead of calling CreateImageTemplateCommand and
CloneSingleCinderDiskCommand directly.

AddVmTemplateFromSnapshotCommand runs
CreateAllTemplateDisksFromSnapshotCommand for the same purpose.

CreateAllTemplateDisksCommand incorporates callback that is needed to
poll the child commands that create disk images. The callback in
AddVmTemplateCommand orchestrates the process of template creation,
including post-processing of template disk images, which will be added
in the next patches.

Change-Id: I6fc7e46df640a5ec0c0c7bc0e3f512e16d6969ae
Bug-Url: https://bugzilla.redhat.com/1436219
Signed-off-by: Shmuel Melamud <smelamud@redhat.com>