core: StorageJobCallback - adding support for child commands
This patch adds support to the StorageJobCallback for executing and
monitoring child commands.
In order to do so, the StorageJobCallback was modified to extend
ChildCommandsCallbackBase (which provides the child commands monitoring
support).

This change is needed in order to add support for storage job fencing,
which is supposed to be performed by child commands while polling
storage job.

Change-Id: I8fa37385bad0322de5ad07623e951be28284045a
Bug-Url: http://bugzilla.redhat.com/1409125
Signed-off-by: Liron Aravot <laravot@redhat.com>