core: StorageJobs - adding support for job fencing
This patch adds support for storage job fencing for entity polling
commands in the StorageJobCallback.

When a job possibly hasn't started yet, we can attempt to fence the job
using the entity that may be used to poll the job  status. Fencing the job
means that the host that recieved it will fail when attempting to
execute it.
Fencing the jobb would usually be performed by executing an asynchronous
fencing command on the entity the job is supposed to be performed on.
Once the job was fenced, the entity poller in use should be able to determine
the job status as failed.

Change-Id: Iaf318884f0d9667d80eaff7f31bf4fdd39a6afb1
Bug-Url: http://bugzilla.redhat.com/1409125
Signed-off-by: Liron Aravot <laravot@redhat.com>