core: adding support for fencing storage volume job
This patch adds the infrastracture required to fence volume related
jobs.
In case the host the job was sent to becomes non responsive before the
job was started, the engine may attempt to fence the operation using the
volume the job was supposed was performed on.

Fencing a volume job is done by incrementing the volume generation by a
predefined interval using the FenceVolumeJob command. If the fencing
operation succeeded, it'll be detect by the VdsmImagePoller on the next
polling attempt of the StorageJobCallback.

The next patch in the series adds use of the added infrastructure in the
CopyData command.

Change-Id: Ib51944d0ab3e16c837bb117589e8e22f9756b219
Bug-Url: http://bugzilla.redhat.com/1409125
Signed-off-by: Liron Aravot <laravot@redhat.com>