core: VdsmImagePoller - ILLEGAL status consideration
This patch modifies the VdsmImagePoller to not consider the ILLEGAL
status for determining the result of a volume job.

As the SDM.update_volume verb was added, the rule of thumb that if a
volume is ILLEGAL the operation has failed was changed.

That verb allows changing the volume status to ILLEGAL, which means
that on that case this status doesn't indicate on failure.

This patch changes the VdsmImagePoller to fit to that new case, when the
lease is FREE the poller will check the volume generation -
a. if the generation equals to the execution generation + 1 the job
succeeded.
b. if the generation equals to the execution generation + the fencing
diff the operation can be considered as failed.
c. on any other case, the job status can't be determined by the poller.

when the generation equals to the execution generation - the job may be
pending or it may never start if the host was shut down. The poller will
wait till the volume job fencing will succeed or until the operation will
end) and then will end the job according to checks a/b.

Change-Id: I9794cb7ba1259791ed4c7d699fabdeae88b21cbf
Bug-Url: http://bugzilla.redhat.com/1409125
Signed-off-by: Liron Aravot <laravot@redhat.com>