backend: added isIpv6MigrationProperlyHandled to FeatureSupported
ipv6 migration/literal ip addressess are due to be introduced in 4.2,
where vdsm properly handles/ensures ipv6 literal addresses to be used
in qemu uri's.

until then, a temporary solution that simply wraps ipv6 addresses in
square brackets (in accordance with rfc2732) is planned to be placed
on top of this patch.

Change-Id: I7cafa52dca0b3364efc560319ed668a7a6122d90
Signed-off-by: Leon Goldberg <lgoldber@redhat.com>