core: better support in the monitoring for auto start of vms with lease
Prevent the VM from switching to Down state upon failure to start on a host
while being in Unknown state on another host:
1. We do not change the status and run_on_vds fields of the VM on run VM.
2. Clear the async running VM operation before auto start attempt to avoid
errors of type 'VM is during initiating'.
3. Clear the async running VM when a VM is restored from Unknown status
(but not when it goes to WaitForLaunch or MigratingTo statuses that do
not necessarily mean that the VM is actually running).

Change-Id: Icca6920dce600959045d8885404a0de056b88c8f
Bug-Url: https://bugzilla.redhat.com/1317429
Signed-off-by: Arik Hadas <ahadas@redhat.com>