core: TransferImage - callback end methods NPE
When the TransferImageCommand is being executed for pre existing disk
it'll attempt to load the disk from the db on the OnSucceeded/OnFailed
methods using the image id provided in the command parameters while for a
created disk by the command the data will be taken from the AddDisk parameters.

When uploading to a pre existing disk, the id provided by the user
is passed the in the parameters image id (field which should represent a volume)
instead of in the image group id field (which should represnt the disk id).
When attempting to load the image in the end method using getDiskImage()
no data will be returned as the disk id is used as the volume id and
NPE will be thrown.
In other parts of the command the disk is loaded manually, therefore
this exception is thrown only on the end methods.

Looking forward, we should pass the id provided by the user in the
parameters image group id and complete the image id from the db - this
patch fixes the problem by altering the log in order to reduce the risk
in backporting this change to the stable branches.

Change-Id: Ice26f4d19723042097a81ea4b685a7aa31faf976
Bug-url: https://bugzilla.redhat.com/1400299
Signed-off-by: Liron Aravot <laravot@redhat.com>