api: DisksResource - initial size support on add disk
When a sparse image disk is created on block domain it's created with
initial allocation and being extended as needed by the spm (when it's
        plugged to a running vm).
Speicifying inital size on the disk creation is useful when there's a
need to create the disk with initial allocated different than the
default value (1gb) for various reasons.
An example would be the image upload, when the data is uploaded to disk
extensions aren't supported - so the disk needs to be created with
initial size that is big enough to contain all the uploaded data.

Change-Id: I463c46a0072076fb0db8ba78a3b587ca34868771
Bug-Url: https://bugzilla.redhat.com/1400296
Signed-off-by: Liron Aravot <laravot@redhat.com>