backend: move illegal Pass Discard logging to AddVmCommand
This patch moves the logging of disks with illegal Pass Discard value
from AddVmFromTemplateCommand to AddVmCommand.
Therefore, the log will not be shown only when creating a cloned vm from
a template, but also when creating a thin vm from a template, when
cloning a vm and when adding a vm from a snapshot.

Change-Id: I84349befab5c9752825b0667a327edddae1d3819
Bug-Url: https://bugzilla.redhat.com/1410019
Signed-off-by: Idan Shaby <ishaby@redhat.com>