backend: handle video devices for vm next run
In case of updating a running VM, the createNextRunSnapshot() updates
the "Number of monitors" property in ovf only if at least one video
device exists for the VM.

Till now we always removed the video devices in
VmDeviceUtils.getVmDevicesForNextRun() - for both Headless and
non Headless VMs (since the video device will be created anyway,
based on the vm.DefaultDislayType value).

But since the "Number of Monitors" based on video device existence, this
patch makes sure such a device exists in case of non headless.

Change-Id: Ifb6e3719ccd44c8371442f928b934de37ad7ab55
Bug-Url: https://bugzilla.redhat.com/1408691
Signed-off-by: Sharon Gratch <sgratch@redhat.com>