core: Fix bug in discard checks
When passing a list of VM disks to validate if their discard value is
legal a map of a disk to its DiskVmElement is created.
The map is created by fetching all of the VM's DiskVmElements and
iterating on them, however, in some cases not all the VM's disk are sent
for the check (one example is deactivated disks) and the map is created
with a null key (since the corresponding disk does not exist in the list
of disks to check) thus causing an NPE later in the flow.
This was fixed by creating a map entry only if there is a disk in the list
corresponding to the DiskVmElement fetched

Change-Id: I5379f3ced6aeeed27e9ab572ff141c0711bd4454
Bug-Url: https://bugzilla.redhat.com/1409995
Signed-off-by: Tal Nisan <tnisan@redhat.com>