core: RegisterVdsQuery RefObject
Use the standard java.util.concurrent.atomic.AtomicBoolean instead of
the deprecated org.ovirt.engine.core.compat.RefObject.
While the atomicity isn't an issue here, AtomicBoolean is a clean and
easy way to get a mutable boolean and to remove another usage of
RefObject, as part of the effort to remove it completely.

Change-Id: I8f8cfee45a2001f84b655288ee15e1cf10312496
Signed-off-by: Allon Mureinik <amureini@redhat.com>