engine: for bonds mode=4 setting default xmit_hash_policy
On creating new bonds with mode=4 (Dynamic link aggregation (802.3ad))
trough the web UI, the  xmit_hash_policy is set explicitly to the
numeric value 2 (layer2+3).

The value is set numerically to be consistent to vdsm, which reports
values of bonding options numerically, instead of textual.

Already available bonds with mode=4 without xmit_hash_policy=2 are not
changed and reported as "custom" options in the web UI.

The previously used default xmit_hash_policy=layer2 uses just
hardware MAC addresses of source and destination to decide which slave
of the bond is used to transmit.
The new default xmit_hash_policy=layer2+3 additionally considers the
layer 3 addresses, if available, and is still 802.3ad compliant.
This policy is intended to provide a more balanced distribution of
traffic than layer2 alone, especially in environments where a layer3
routing is required to reach most destinations.

Change-Id: I5b6cb2dfe8f0b60a1e81aa4832bc8c10672f6bb8
Bug-Url: https://bugzilla.redhat.com/1397490
Signed-off-by: Dominik Holler <dholler@redhat.com>