core: GlusterSyncJobTest argument matchers
While many of the actual underlying implementations of both the test and
the production code pass ArrayLists, this is not required by neither the
methods' contracts, or their internal logic. Having the matchers
explicitly verify that ArrayLists are passed (instead of just
Collections or Lists, as specified by the methods' contracts) poses an
unwarranted strictness on the test, causing it to test implementation
details instead of the contractual behavior, and potentially
[erroneously] fail if the underlying implementation is changed to use
some other, potentially better suited collection.

Change-Id: I3bd92ad97025390862d7636a8e009cc1f75d7fe1
Signed-off-by: Allon Mureinik <amureini@redhat.com>