ImageTransfer - removing image link
We currently use "image" to pass the disk id when performing image
transfer - but currently we don't support image links which causes the
link to appear as "/null".

Looking forward, we'll replace "image" with "disk" which will let us
to provide a link to the disk which is used in the transfer.

Change-Id: Iba9a10c4e153381753e1f66a69edfb23ee7e4c20
Bug-Url: https://bugzilla.redhat.com/1401873
Signed-off-by: Liron Aravot <laravot@redhat.com>