core: fix NumberFormatException when importing a removed VM
When importing a VM while another is in the process of being
removed a NumberFormatException will be thrown and the import
will fail.

This is caused by the removed VMs getting a "_remove_me_"
prefix to its GUID when removed. Because of the prefix the
method GetImagesListVDSCommand#executeIrsBrokerCommand fails
when attempting to create an instance of the Guid object
using the prefixed GUID of the removed VM.

Fixed by filtering out the prefixed image GUIDs in
ImagesListReturn constructor.

Change-Id: I564fb9d0c08e6fabbf8ffd05cfdf93d5435a0e0b
Bug-Url: https://bugzilla.redhat.com/1331335
Signed-off-by: Benny Zlotnik <bzlotnik@redhat.com>