core: Fix HotUnplugMemorySupported definition
HotUnplugMemorySupported is disabled in the code and so far it was not
enabled in any version, so there's no need to have the value in db and
we can rely on the default defined in ConfigValues.

When the feature becomes enabled in future, we will update values in db
accordingly.

Change-Id: I3b3d042dc4d331f11272cbbbf23eacf5e09004dd
Signed-off-by: Martin Perina <mperina@redhat.com>