core: AddDiskCommand VM existence validation
Commit 801c1c introduce a subtle bug where parameters containing a
valid VmDiskElement object that points to a non-existent VM would result
in the command passing its validate() method, but adding a validation
error, which is an illegal combination (not to mention a logically wrong
one - adding a disk to a non-existent VM should just fail).

This patch fixes the bug in the production code and updates the test
accordingly to mock the VM where needed.

Change-Id: I32036397a5718d1833646c96a8dbf7aab59b5171
Signed-off-by: Allon Mureinik <amureini@redhat.com>